% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_measures.R
\name{approx_angle}
\alias{approx_angle}
\title{Calculate angle between geometries}
\usage{
approx_angle(x, y)
}
\arguments{
\item{x, y}{Geometry vectors. Only surfaces, curves, and arrows are allowed}
}
\value{
A numeric vector given angular difference in radians
}
\description{
Angles cannot be given exactly since vector angle relies on the vector length
as well as \link{acos}. This function calculate the angle between two geometries.
This is defined for surfaces, curves, and arrows. If a surface is supplied
the angle will be calculated based on the surface normal and modified to
match the original geometry.
}
\examples{
# Angle between two lines
approx_angle(line(3, 7, -1), line(-5, 4, 2))

# Angle between vector and plane
approx_angle(plane(-4, 1, 19, -4), vec(5, -7, 2))

}
