% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{bisector}
\alias{bisector}
\title{Calculate the bisector of two geometries}
\usage{
bisector(x, y)
}
\arguments{
\item{x, y}{Vector of points}
}
\value{
A vector of lines or planes depending on the dimensionality of x and
y
}
\description{
A bisector is a geometry that splits the space in two, i.e. a line in 2D and
a plane in 3D. Constructing a bisector from 2 geometries creates a split
where the edge at any given point has equal length to the two geometries.
}
\examples{
bisector(point(4, -2), point(7, 5))

}
