% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_boolop.R
\docType{import}
\name{boolean_operations}
\alias{boolean_operations}
\alias{reexports}
\alias{union}
\alias{difference}
\alias{symmetric_difference}
\alias{complement}
\title{Boolean operations on geometries}
\usage{
difference(x, y, ...)

symmetric_difference(x, y, ...)

complement(x, ...)
}
\arguments{
\item{x, y}{vectors of geometries}

\item{...}{arguments passed on to methods}
}
\value{
Return value will depend on the different implementations
}
\description{
While intersection calculation and queries are the only part of the boolean
operators provided by euclid, it provides generics for the remaining ones so
that other packages may fill the gap (e.g. polyclid for 2D boolean operations
on polygons).
}
\seealso{
Other Boolean operations: 
\code{\link{intersection}()}
}
\concept{Boolean operations}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:setops]{union}}}
}}

