% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{centroid}
\alias{centroid}
\title{Calculate the centroid of a geometry or a set of points}
\usage{
centroid(x, y = NULL, z = NULL, t = NULL)
}
\arguments{
\item{x}{A vector of triangles or tetrahedrons, or points if \code{y}, \code{z}, or \code{t}
are provided}

\item{y, z, t}{Vector of points}
}
\value{
a vector of points
}
\description{
The centroid is the geometric mean of the vertices in a geometry. euclid
supports centroid calculations for triangles, tetrahedrons and 2-4 points
}
\examples{
p <- point(c(1, 5, 6), c(1, -2, 4))
# centroid of triangle
t <- triangle(p[1], p[2], p[3])
centroid(t)

plot(t, col = "grey", border = NA)
euclid_plot(centroid(t), pch = 16, cex = 2, col = "firebrick")

# Same as providing the points directly
centroid(p[1], p[2], p[3])

}
