% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{chart}
\alias{chart}
\title{Create a chart for thr given objects}
\usage{
chart(..., scale = TRUE, asp = 1, axes = TRUE, ylab = NA, xlab = NA,
  par = NULL)
}
\description{
Create a chart for thr given objects
}
