% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{circumcenter}
\alias{circumcenter}
\title{Calculate circumcenter of triangles, tetrahedrons, or sets of points}
\usage{
circumcenter(...)
}
\arguments{
\item{...}{geometry vectors. Either a triangle or tetrahedron vector, or 2,
3, or 4 point vectors. If 4 are provided they must be in 3D}
}
\value{
A points vector
}
\description{
The circumcenter is the center of the circle of sphere passing through the
vertices of a geometry (or the provided points).
}
\examples{
# Circumcenter of triangle
t <- triangle(point(1, 1), point(5, -2), point(6, 4))
circumcenter(t)

# Same as providing the points directly
circumcenter(point(1, 1), point(5, -2), point(6, 4))

# Same as getting the center of a circle constructed from the points
circ <- circle(point(1, 1), point(5, -2), point(6, 4))
vertex(circ)

plot(circ)
euclid_plot(t, col = "grey", border = NA)
euclid_plot(circumcenter(t), cex = 3, col = "firebrick", lwd = 2)
euclid_plot(vertex(circ), cex = 1.5, col = "steelblue", lwd = 2)

}
