% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{clearpolys}
\alias{clearpolys}
\title{Push m in the direction of v until m \cap X = \emptyset}
\usage{
clearpolys(m, X, v = C(m), vector = FALSE)
}
\description{
Push m in the direction of v until m \cap X = \emptyset
}
