% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_predicates.R
\name{collinear}
\alias{collinear}
\alias{coplanar}
\title{Check geometries for whether they are collinear or coplanar}
\usage{
collinear(x, y, z = NULL)

coplanar(x, y, z = NULL, t = NULL)
}
\arguments{
\item{x, y, z, t}{Vector of geometries.}
}
\value{
A logical vector
}
\description{
Geometries are collinear if they lie on the same line and coplanar if they
lie on the same plane. collinearity can be tested between two curves or three
points, weighted points, circles, or spheres. coplanarity can be tested
between two surfaces or 4 points, weighted points, or spheres
}
\examples{
# Collinear points
collinear(point(0, 0), point(1, 1), point(2, 2))

# Coplanar triangles
coplanar(
  triangle(point(0, 0, 0), point(1, 1, 1), point(1, 1, 0)),
  triangle(point(0, 0, 0), point(-1, -1, -1), point(-1, -1, 0))
)

}
