% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{compose}
\alias{compose}
\title{Arranges shapes in m on layout in km according to membs vector}
\usage{
compose(m, shp, km, noverlap = noverlap_radial, ...)
}
\description{
Arranges shapes in m on layout in km according to membs vector
}
