% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_predicates.R
\name{constant_in}
\alias{constant_in}
\alias{is_constant_in}
\alias{has_constant_x}
\alias{has_constant_y}
\alias{has_constant_z}
\title{Check geometries for axis alignment}
\usage{
is_constant_in(x, axis)

has_constant_x(x)

has_constant_y(x)

has_constant_z(x)
}
\arguments{
\item{x}{A geometry vector}

\item{axis}{One or more specifications of the axes to check against as \code{"x"},
\code{"y"}, \code{"z"} or \code{1}, \code{2}, \code{3}}
}
\value{
A logical vector
}
\description{
These predicate functions checks for whether a geometry varies along a given
axis. While it is common in 2 dimensions to ask whether something is vertical
or horizontal that notion does not scale well to 3 dimensions and euclid
instead elects to check for whether a given geometry is constant in a
specific coordinate value. As such, e.g. \code{has_constant_x()} is equivalent to
asking whether a given 2 dimensional geometry is vertical but also works for
3 dimensional geometries.
}
\examples{
# Check for horizontal vector
v <- vec(1, -2:2)
has_constant_y(v)

# Use recycling to check all axes
is_constant_in(vec(2, 0, -5), c("x", "y", "z"))

}
