% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{Access coordinate dimensions in geometries}
\usage{
coord(x, dimension)
}
\arguments{
\item{x}{A geometry vector}

\item{dimension}{The dimension to retrieve coordinate value from, either as
a character (\code{"x"}, \code{"y"}, or \code{"z"}), or an integer.}
}
\value{
An exact_numeric vector
}
\description{
This method gives access to the coordinates of geometries (if defined) as
\link{exact_numeric} vectors.
}
\examples{
points <- point(1:10, 10:1)
coord(points, "x")

}
