% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\name{def}
\alias{def}
\alias{def.euclid_geometry}
\alias{def.euclid_affine_transformation}
\alias{def<-}
\alias{def<-.euclid_geometry}
\alias{def<-.euclid_affine_transformation}
\alias{definition_names}
\title{Access the exact values that defines the geometries}
\usage{
def(x, ...)

\method{def}{euclid_geometry}(x, name, which = NULL, ...)

\method{def}{euclid_affine_transformation}(x, i, j, ...)

def(x, ...) <- value

\method{def}{euclid_geometry}(x, name, which = NULL, ...) <- value

\method{def}{euclid_affine_transformation}(x, i, j, ...) <- value

definition_names(x, ...)
}
\arguments{
\item{x}{A geometry vector}

\item{...}{parameters to pass on}

\item{name}{Either a name or the index of the definition to extract, as
matched to \code{definition_names(x)}}

\item{which}{An integer vector giving the vertex from which the definition
should be extracted, or \code{NULL} to extract all
to extract all.}

\item{i, j}{The row and column of the cell in the transformation to fetch.}

\item{value}{An \code{exact_numeric} vector or an object convertible to one}
}
\value{
An exact_numeric vector
}
\description{
This function gives access to the underlying values defining the geometries
in a vector. As such they return the same information as calling
\code{\link[=as.matrix]{as.matrix()}} on a geometry vector except the return value is kept as an
exact numeric and that you can extract from single elements if the
cardinality of the geometry exceeds 1.
}
\examples{
# Get squared radius of circle
circ <- circle(point(4, 7), 25)
def(circ, "r2")

# Set r2 to 10
def(circ, "r2") <- 10
circ

# Get all the x values from the source of segments
s <- segment(point(sample(10, 4), sample(10, 4)),
             point(sample(10, 4), sample(10, 4)))
def(s, "x", 1L)

# Get y for all subelements
def(s, "y")

# Extract cell values from transformation matrices
m <- affine_rotate(c(pi/2, pi/3))
def(m, 1, 2)

}
\seealso{
Other Geometry methods: 
\code{\link{euclid_geometry}},
\code{\link{subgeometries}}
}
\concept{Geometry methods}
