% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\alias{is_direction}
\alias{as_direction}
\title{Vector of directions}
\usage{
direction(..., default_dim = 2)

is_direction(x)

as_direction(x)
}
\description{
Vector of directions
}
