% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_measures.R
\name{distance_squared}
\alias{distance_squared}
\alias{approx_distance_matrix}
\title{Calculate distances between geometries}
\usage{
distance_squared(x, y)

approx_distance_matrix(x, y)
}
\arguments{
\item{x, y}{eometry vectors or bounding boxes}
}
\value{
A \code{euclid_exact_numeric} vector for \code{distance_squared()} and a
numeric matrix for \code{distance_matrix()}
}
\description{
The minimum distance between two arbitrary geometries is non-trivial and is
only exactly defined for non-circular geometries. \code{distance_squared()} will
return the exact squared distance between geometries with \code{x} and \code{y} being
recycled to the maximum length of either. \code{distance_matrix} will return a
matrix of distances given as numerics (and thus not exact), with the
geometries of \code{x} in the rows and the geometries of \code{y} in the columns so
that the value of \code{mat[i, j]} corresponds to the distance between \code{x[i]} and
\code{y[j]}.
}
\examples{
# Calculate distances between lines and rays in 3D
p <- point(sample(100, 20), sample(100, 20), sample(100, 20))
l <- line(p[1:5], p[6:10])
r <- ray(p[11:15], p[16:20])

# Pairwise exact distance
distance_squared(l, r)

# All distances
approx_distance_matrix(l, r)

}
