% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{equidistant_line}
\alias{equidistant_line}
\title{Construct the equidistant line of 2 or three points}
\usage{
equidistant_line(x, y, z = NULL)
}
\arguments{
\item{x, y, z}{Vector of points. If z is given they must be in 3D, otherwise in
2D}
}
\value{
A vector of lines
}
\description{
An equidistant line is a line that at any given point is the same distance to
the points that defines it. In 2 dimensions it is equivalent to the bisector
of the two points. In 3 dimensions it is defined for 3 points (2 points would
give a bisector plane).
}
\examples{
equidistant_line(
  point(3, 7, 2),
  point(6, -4, 9),
  point(2, 1, -5)
)

# In 2D you get the bisector
equidistant_line(point(2, 5), point(3, 2)) == bisector(point(2, 5), point(3, 2))

}
