% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid-package.R
\docType{package}
\name{euclid-package}
\alias{euclid}
\alias{euclid-package}
\title{euclid: Exact Computation Geometry Framework Based on 'CGAL'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a rich framework for doing computation geometry by 
    interfacing with the 'CGAL' library. The 'euclid' package provides new 
    vector types for standard geometric objects and supports exact computations
    on them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/euclid}
  \item Report bugs at \url{https://github.com/thomasp85/euclid/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
