% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid-package.R, R/geometry_intersection.R,
%   R/geometry_locations.R, R/geometry_measures.R
\name{euclid_extend}
\alias{euclid_extend}
\alias{intersection_impl}
\alias{has_intersection_impl}
\alias{centroid_impl}
\alias{distance_squared_impl}
\alias{approx_distance_matrix_impl}
\title{Extension points}
\usage{
intersection_impl(x, y)

has_intersection_impl(x, y)

centroid_impl(x, ...)

distance_squared_impl(x, y)

approx_distance_matrix_impl(x, y)
}
\arguments{
\item{x, y}{vectors}

\item{...}{additional arguments}
}
\value{
Depends on function. Consult the main function to see the expected
output (e.g. \code{\link[=centroid]{centroid()}} for \code{centroid_impl()})
}
\description{
Euclid is build to be extended. Some of the core functions in the package are
generics that other packages can provide methods to, but in some
circumstances it was easier to provide a more indirect extension mechanism
by providing methods for one or more of the generics below:
}
\keyword{internal}
