% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidist.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean_distance}
\usage{
euclidean_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1:}{x coordinate of the first point}

\item{y1:}{y coordinate of the first point}

\item{x2:}{x coordinate of the second point}

\item{y2:}{y coordinate of the second point}
}
\value{
distance: euclidean distance between the two points
}
\description{
Compute the euclidean distance between two points
}
\examples{
euclidean_distance(x1=2, y1=3, x2=2, y2=2)
}
\author{
Benedetta Bianchi
}
