% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_predicates.R
\name{geometry_class}
\alias{geometry_class}
\alias{is_surface}
\alias{is_volume}
\alias{is_curve}
\title{Query geometry class}
\usage{
is_surface(x)

is_volume(x)

is_curve(x)
}
\arguments{
\item{x}{A geometry vector}
}
\value{
a logical scalar
}
\description{
While the geometry types provided by euclid are primitives, they can be
grouped based on their properties. Lines, rays, and segments are all curves
in space, with rays and segments being the semi-finite and finite version of
lines respectively. Circles, planes, triangles, and iso rectangles are all
surfaces in the sense that they define an area but doesn't have a volume.
Spheres and iso cubes are volumes. The class of geometry can be checked with
the given functions
}
\examples{
l <- line(3, 8, 1)
is_surface(l)
is_curve(l)
is_volume(l)

}
