% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_measures.R
\name{geometry_measures}
\alias{geometry_measures}
\alias{approx_length}
\alias{approx_area}
\alias{approx_volume}
\alias{approx_radius}
\title{Approximate geometry measures}
\usage{
approx_length(x)

approx_area(x)

approx_volume(x)

approx_radius(x)
}
\arguments{
\item{x}{A geometry vector}
}
\value{
A numeric vector
}
\description{
These functions provide approximate measures of length, area, and volume of
geometries where relevant. The reason for approximation is that some measures
require square root operations or multiplication by π, both operations where
exactness is necessarily lost. Not all geometries have meaningful measures,
e.g. a direction is dimensionless, a line is without area, and a circle has
no volume. Some geometries are infinite in some measures, e.g. a plane has an
infinite area. The return value will reflect this.
}
