% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_projection.R
\name{geometry_projection}
\alias{geometry_projection}
\alias{project}
\alias{map_to}
\title{Project geometries to lines and planes}
\usage{
project(x, target)

map_to(x, target)
}
\arguments{
\item{x}{A vector of geometries to project or map}

\item{target}{A vector of lines (for projection) or planes (for projection or
mapping)}
}
\value{
A vector of geometries. For projection the class match that of \code{x},
for mapping it is the 2 dimensional counterpart to \code{x}
}
\description{
These functions handles projections and mappings of geometries to lines and
planes. For projections, the geometry type will stay the same, while for
mapping the geometries will be converted to their 2D variants. Not all
geometries can be projected and mapped. Specifically, iso rectangles and iso
cuboids cannot do either, only 3 dimensional geometries can be projected to a
plane, and only 3 dimensional geometries with a 2 dimensional counterpart can
be mapped to a plane. Circles and spheres are special in the sense that no
ellipse and ellipsoid primitives exist. Because of this, projections will
retain the radius and circularity of the primitives despite this not being
geometrically sound. Projecting a 3D circle to a line will keep the
orientation of the circle, while projecting it to a plane will adopt the
orientation of the plane. Mapping a 2D circle to a plane will map the center
and keep the radius. Mapping a sphere to a plane will produce a circle.
}
