% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_intersection.R
\name{has_intersection}
\alias{has_intersection}
\alias{\%is_intersecting\%}
\title{Query whether geometries intersect}
\usage{
has_intersection(x, y)

x \%is_intersecting\% y
}
\arguments{
\item{x, y}{Geometry vectors or bounding boxes}
}
\value{
a logical vector
}
\description{
While it is sometimes necessary to get the actual intersection geometry, it
may other times only be necessary to query whether an intersection exists.
This is in general much faster than checking if \code{\link[=intersection]{intersection()}} returns
\code{NULL}. Further, it is not bounded by the same constraint that an exact
intersection must be constructable so a wider range of geometries are
supported for the predicate. If the query is not possible it will return \code{NA}
rather than an error
}
\examples{
# Example of the predicate
t <- triangle(point(0, 0), point(1, 1), point(0, 1))
l <- line(1, -1, c(0, 1, 2))
has_intersection(t, l)

# Use binary operator instead
t \%is_intersecting\% l

# 2 dimensional circles are better supported
l \%is_intersecting\% circle(point(7, 4), 9)

}
