% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_predicates.R
\name{is_degenerate}
\alias{is_degenerate}
\title{Check if a geometry is degenerate}
\usage{
is_degenerate(x)
}
\arguments{
\item{x}{A geometry vector}
}
\value{
A logical vector
}
\description{
Geometries can be ill-defined for several reasons, e.g. circles with a radius
of zero, segments that start and end at the same location, etc. See details
for a list of conditions that are considered degenerate.
}
\details{
The following conditions are considered to produce degenerate geometries:
\itemize{
\item A circle having a radius of zero
\item An iso-rectangle with zero width or zero height
\item An iso-cuboid with zero width, zero height, or zero depth
\item A 2D line with the coefficients being zero
\item A 3D line that is degenerated to a point
\item A plane with the coefficients being zero
\item A ray with the source and the second defining point being equal
\item A segment with the source and target being equal
\item A sphere with a radius of zero
\item A tetrahedron with all vertices being co-planar
\item A triangle with collinear vertices
}
}
\examples{
# A line without direction
l <- line(point(0, 5), vec(0, 0))
is_degenerate(l)

# A circle without radius
circ <- circle(p(3, -1), 0)
is_degenerate(circ)

}
