% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{line}
\alias{line}
\alias{is_line}
\alias{as_line}
\title{Vector of lines}
\usage{
line(..., default_dim = 2)

is_line(x)

as_line(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A vector of lines or an object to convert to it}
}
\value{
An \code{euclid_line} vector
}
\description{
A line is an undirected infinite line. For 2D it can be defined by the line
equation \code{ax + by + c = 0} whereas for 3D it is usually defined by a point
and a direction in the same way as rays, but implicitly extending to infinity
in the opposite direction as well.
}
\section{Constructors}{

\strong{2 dimensional rays}
\itemize{
\item Providing 3 numerics will create lines with the given line equation
\item Providing two points will construct lines going through those
\item Providing a point and a vector will construct lines going through the point
and extending in the direction of the vector.
\item Providing a ray will construct the supporting line for the ray
\item Providing a segment will construct the supporting line for the segment
}

\strong{3 dimensional rays}
\itemize{
\item Providing two points will construct lines going through those
\item Providing a point and a vector will construct lines going through the point
and extending in the direction of the vector.
\item Providing a ray will construct the supporting line for the ray
\item Providing a segment will construct the supporting line for the segment
}
}

