% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_projection.R
\name{map_to}
\alias{map_to}
\title{Map 3D geometries to 2D based on plane}
\usage{
map_to(x, target)
}
\arguments{
\item{x}{A vector of geometries to project or map}

\item{target}{A vector of planes}
}
\value{
A vector of geometries in 2 dimensions
}
\description{
3 dimensional geometries can be converted to 2 dimensions by mapping them to
a plane. Not all 3 dimensional geometries can be converted to 2 dimensions as
it requires for them to have a 2 dimensional counterpart. This rules out iso
cubes, planes, and tetrahedrons. Spheres are a special case that will get
converted to circles with the same radius. As no ellipsoid geometry exist
the plane of circles is ignored and mapping a 3D circles thus creates a 2D
circle of the same radius irrespective of the supporting plane of the circle.
}
\examples{
# Map a 3D point cloud to a plane defined by three random points
p <- point(sample(10), sample(10), sample(10))
support <- p[sample(10, 3)]
map_to(p, plane(support[1], support[2], support[3]))

}
