% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{p_is_ccw}
\alias{p_is_ccw}
\alias{p_is_ccw}
\title{Is polygon P stored in counterclockwise order?}
\usage{
p_is_ccw(P)

p_is_ccw(P)
}
\description{
Is polygon P stored in counterclockwise order?

Is polygon P stored in counterclockwise or clockwise order?
TODO: generalize to arbitrary planes
}
