% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\alias{parameter.euclid_geometry}
\alias{parameter.euclid_affine_transformation}
\title{Access the exact values that defines the geometries}
\usage{
parameter(x, ...)

\method{parameter}{euclid_geometry}(x, which, element = NA, ...)

\method{parameter}{euclid_affine_transformation}(x, i, j, ...)
}
\arguments{
\item{x}{A geometry vector}

\item{...}{parameters to pass on}

\item{which}{Either a name or the index of the definition to extract, as
matched to \code{definition_names(x)}}

\item{element}{For geometries with a cardinality above 1, which element of
the geometry should the definition be extracted for. If \code{NA} the definition
for all elements will be returned and the length of the returned vector will
be \code{sum(cardinality(x))} (matching the return of \code{as.matrix(x)})}

\item{i, j}{The row and column of the cell in the transformation to fetch.}
}
\value{
An exact_numeric vector
}
\description{
This function gives access to the underlying values defining the geometries
in a vector. As such they return the same information as calling
\code{\link[=as.matrix]{as.matrix()}} on a geometry vector except the return value is kept as an
exact numeric and that you can extract from single elements if the
cardinality of the geometry exceeds 1.
}
