% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\alias{is_point}
\title{Vector of points}
\usage{
point(x, y, z = NULL)

is_point(x)
}
\arguments{
\item{x, y, z}{\link{exact_numeric} vectors or standard R numerics. If \code{z} is \code{NULL}
a 2-dimensional point is constructed, if not a 3 dimensional point is
constructed.}
}
\value{
a \code{euclid_point2} or \code{euclid_point3} vector
}
\description{
Points are the fundamental unit in geometry, from which other geometric
primitives can be constructed.
}
\examples{
num1 <- exact_numeric(runif(5))
num2 <- exact_numeric(runif(5))
num3 <- exact_numeric(runif(5))

# 2 dimensions
point(num1, num2)

# 3 dimensions
point(num1, num2, num3)

# Standard R vectors are automatically converted to exact_numeric vectors
point(runif(5), runif(5))

}
