% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{position}
\alias{position}
\title{Position object by translation and rotation}
\usage{
position(m, v = vdif(O(dct(m)), C(m)), t = theta(v, vhead(m)))
}
\description{
Position object by translation and rotation
}
