% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{pp_clear_all}
\alias{pp_clear_all}
\title{Push P in the direction of v until clear of all shapes in m}
\usage{
pp_clear_all(m0, m1, km, v = C(m0), value = c("s", "v", "p"),
  aprox = FALSE, sort = NULL)
}
\description{
Push P in the direction of v until clear of all shapes in m
}
