% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{pp_clear_poly}
\alias{pp_clear_poly}
\title{Push m along v until completely clear of X}
\usage{
pp_clear_poly(m0, X, v = C(m), value = c("s", "v", "p"), aprox = FALSE)
}
\description{
Push m along v until completely clear of X
}
