% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_projection.R
\name{project}
\alias{project}
\title{Project geometries to lines and planes}
\usage{
project(x, target)
}
\arguments{
\item{x}{A vector of geometries to project or map}

\item{target}{A vector of lines or planes}
}
\value{
A vector of geometries of the same class as \code{x}
}
\description{
Projecting geometries to lines and planes transforms the geometries to lie on
the given line or plane by movement orthogonal to the line or plane. The
geometry of the result of the projection is the same as the input. Not all
geometries can be projected to lines and planes. Specifically, iso rectangles
and iso cubes cannot do either and only 3 dimensional geometries can be
projected to a plane. Circles and spheres are special in the sense that no
ellipse and ellipsoid primitives exist. Because of this, projections will
retain the radius and circularity of the primitives despite this not being
geometrically sound. Projecting a 3D circle to a line will keep the
orientation of the circle, while projecting it to a plane will adopt the
orientation of the plane.
}
\examples{
# Project a collection of segments to a line
p <- point(sample(10), sample(10))
s <- segment(p[1:5], p[6:10])

project(s, line(3, 1, -6))

# Project a line to the x plane
project(line(point(3, 7, -3), point(-2, 6, 0)),
        plane(point(0, 0, 0), vec(1, 0, 0)))

}
