% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{radical}
\alias{radical}
\title{Calculate radical line of two circles or radical plane of two spheres}
\usage{
radical(x, y)
}
\arguments{
\item{x, y}{vector of circles in 2D or spheres}
}
\value{
A vector of lines or planes depending on the input
}
\description{
The radical line or plane is the line (or plane) perpendicular to the line
connecting the two circles or spheres and positioned such that at any point
the tangent lines drawn to the two circles or spheres will have equal length.
If the circles or sphere cross the radical will be positioned at the crossing
points. It is not defined for cocentric circles and spheres.
}
\examples{
# Radical line
radical(
  circle(point(3, 6), 12),
  circle(point(-6, 2), 3)
)

# Radical plane
radical(
  sphere(point(3, 6, 3), 12),
  sphere(point(-6, 2, 7), 3)
)

}
