% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random}
\alias{random}
\title{Create n random elements of dimensionality d with the given distribution P}
\usage{
random(n = 10, d = 2, P = function(n) runif(n, -1, 1))
}
\description{
Create n random elements of dimensionality d with the given distribution P
}
