% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray.R, R/segment.R
\name{ray}
\alias{ray}
\alias{is_ray}
\alias{as_ray}
\alias{is_segment}
\title{Vector of rays}
\usage{
ray(..., default_dim = 2)

is_ray(x)

as_ray(x)

is_segment(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A vector of rays or an object to convert to it}
}
\value{
An \code{euclid_ray} vector
}
\description{
A ray is a directed line that starts at a given point and extends to
infinity. As such it does not have a magnitude (like a vector) but only a
beginning and a direction. A ray is considered to be
\link[=is_degenerate]{degenerate} if constructed be two points that are equal (and
thus have no direction). Rays can be flipped by taking their negative
}
\section{Constructors}{

\strong{2 and 3 dimensional rays}
\itemize{
\item Providing two points will construct rays starting at the first point and
going through the second
\item Providing a point and a vector will construct rays starting at the point
and moving in the direction of the vector.
\item Providing a point and a direction will construct rays starting at the point
and moving in the given direction.
\item Providing a point and a line will construct rays starting at the point
and moving in the direction of the line. Note that the point does not have
lie on line.
}
}

\examples{
# Construction
p <- point(sample(10, 3), sample(10, 3))
r <- ray(p[1], p[2:3])
r
plot(p, pch = c(16, 17, 17))
euclid_plot(r)

ray(p, -as_vec(p))

ray(p, direction(7, -2))

ray(p, line(3, 10, -8))

# Flip direction
-r

# Convert to vector, direction, or lines
as_vec(r)

as_direction(r)

as_line(r)

}
