% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_workers.R
\name{s_v_op}
\alias{s_v_op}
\title{Low-level workload functions.
The functions in this file are prime candidates for implementation in Rcpp/RcppParallel/CUDA
Virtually all operations in this package use this functions.
The order of computation is defined in 'storage.R' file.
s -> v}
\usage{
s_v_op(s0, f_, d)
}
\description{
Low-level workload functions.
The functions in this file are prime candidates for implementation in Rcpp/RcppParallel/CUDA
Virtually all operations in this package use this functions.
The order of computation is defined in 'storage.R' file.
s -> v
}
