% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Vector of segments}
\usage{
segment(..., default_dim = 2)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A vector of segments or an object to convert to it}
}
\value{
An \code{euclid_ray} vector
}
\description{
A segment is a finite directed line going from one point to another
}
\section{Constructors}{

\strong{2 and 3 dimensional rays}
\itemize{
\item Providing two points will construct segments starting at the first point
and ending at the second
\item Providing a point and a vector will construct segments starting at the
point and ending at the point defined by the start point plus the vector
}
}

