% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere}
\alias{sphere}
\alias{is_sphere}
\alias{as_sphere}
\title{Vector of spheres}
\usage{
sphere(...)

is_sphere(x)

as_sphere(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{x}{A vector of spheres or an object to convert to it}
}
\value{
An \code{euclid_sphere} vector
}
\description{
A sphere is a 3 dimensional object modelling the surface of a ball. As such
it is an extension of the concept of a 2 dimensional circle to 3D, though a
circle can also exist in three dimensions. A sphere has a center and a
radius. if the radius is 0 it is considered to be \link[=is_degenerate]{degenerate}.
}
\section{Constructors}{

\strong{3 dimensional spheres}
\itemize{
\item Providing 4 points will construct the unique sphere the passes through all
4 points (points must not be coplanar)
\item Providing 3 points will construct the smallest sphere that passes through
all 3 points
\item Providing 2 points will construct the smallest sphere passing through both
points
\item Providing a point and numeric will construct spheres centered on the point
with a squared radius set to the numeric
\item Providing a circle will construct the diametral sphere of the circle
}
}

\examples{
# Construction
p <- point(sample(8), sample(8), sample(8))
sphere(p, 4)

sphere(p[1:2], p[3:4], p[5:6], p[7:8])

sphere(p[1:2], p[3:4], p[5:6])

sphere(p[1:2], p[3:4])

circ <- circle(p[1], as_vec(p[2]), 6)
sphere(circ)

}
