% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{split_sort}
\alias{split_sort}
\title{Split m according to km vector, and list them in the given sort order}
\usage{
split_sort(m, km, sort = NULL)
}
\description{
Split m according to km vector, and list them in the given sort order
}
