% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex.R
\name{subgeometries}
\alias{subgeometries}
\alias{vertices}
\alias{edges}
\alias{n_vertices}
\alias{n_edges}
\title{Vertices and edges of a geometry}
\usage{
vertices(x)

edges(x)

n_vertices(x)

n_edges(x)
}
\arguments{
\item{x}{A vector of geometries}
}
\value{
A \code{euclid_point} or \code{euclid_segment} vector with the requested
geometries.
}
\description{
These functions allows you to extract all the various sub-geometries that a
geometry consists of (if any). \code{vertices()} extracts all the points of each
geometry in a vector and \code{edges()} all the segments. \code{n_vertices()} gives the
number of vertices for each element (equivalent to \code{\link[=cardinality]{cardinality()}}), while
\code{n_edges()} gives the number of edges.
}
\examples{
# not all geometries has both types of subgeometries
rect <- iso_rect(point(1, 1), point(3, 6))

vertices(rect)
edges(rect)
n_vertices(rect)
n_edges(rect)

}
\seealso{
Other Geometry methods: 
\code{\link{def}()},
\code{\link{euclid_geometry}},
\code{\link{vertex}()}
}
\concept{Geometry methods}
