% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron.R
\name{tetrahedron}
\alias{tetrahedron}
\alias{is_tetrahedron}
\alias{as_tetrahedron}
\title{Vector of tetrahedrons}
\usage{
tetrahedron(...)

is_tetrahedron(x)

as_tetrahedron(x)
}
\description{
Vector of tetrahedrons
}
