% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vec}
\alias{vec}
\alias{is_vec}
\alias{as_vec}
\title{Vector of vectors}
\usage{
vec(..., default_dim = 2)

is_vec(x)

as_vec(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A vector of vectors or an object to convert to it}
}
\value{
An \code{euclid_vector} vector
}
\description{
A geometrical vector is somewhat different from the concept of a vector in
programming hence the slightly confusing terminology. In geometry a vector is
a direction and a magnitude most often defined by a point in space where the
direction is defined as the direction from the origin to the point and the
magnitude is defined as the distance from the origin to the point.
}
\section{Constructors}{

\strong{2 dimensional vectors}
\itemize{
\item Providing a point will construct vectors pointing to the points from the
origin centered.
\item Providing two exact numeric vectors will construct vectors pointing to the
point defined by the coordinates given.
}

\strong{3 dimensional vectors}
\itemize{
\item Providing a point will construct vectors pointing to the points from the
origin centered.
\item Providing three exact numeric vectors will construct vectors pointing to
the point defined by the coordinates given.
}
}

