% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex.R
\name{vertex}
\alias{vertex}
\title{Extract points from geometries}
\usage{
vertex(x, which = 1L)
}
\arguments{
\item{x}{A vector of geometries}

\item{which}{An integer vector giving the vertex to extract in cases of
higher cardinality geometries}
}
\value{
A \code{euclid_point} vector matching the dimensionality of x
}
\description{
Geometries located in space (all except directions and vectors), have one or
more points supporting it. Those can be extracted with \code{vertex()}. For
geometries with a cardinality above one there's a choice of which support
point to extract.
}
\section{Vertex definition}{

For geometries that are defined exclusively by points the definition of the
output is straight forward, e.g. for triangles \code{vertex()} will extract one
of the corners depending on the value of \code{which}. For the other geometries
the output is defined according to the below:
\itemize{
\item \strong{circles and spheres}: The vertex is the center
\item \strong{rays}: The vertex is the source
\item \strong{lines and planes}: The vertex is an arbitrary point on the geometry
}
}

\examples{
# Get the source vertex in a segment
s <- segment(point(3, 6), point(1, -7))
vertex(s, 1)

# And the target
vertex(s, 2)

# Arguments are recycled
vertex(s, c(1, 2))

# Get a point on a line
vertex(line(4, 7, -1))

}
