% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_w.R
\name{weighted_point}
\alias{weighted_point}
\alias{is_weighted_point}
\title{Vector of weighted points}
\usage{
weighted_point(..., default_dim = 2)

is_weighted_point(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A point vector or an object to convert to it}
}
\value{
a \code{euclid_point_w} vector
}
\description{
This class is an extension (but not a subclass of) \code{\link[=point]{point()}}. It simply
attaches a numeric weight to the point that is carried along with it.
However, most of the functionality of points are lost with this vector class
as all the a math and sorting operations are gone due to ambiguity on whether
to operate on the point or the weight. Weighted points can easily be
converted to points however.
}
\section{Constructors}{

\strong{2 dimensional weighted points}
\itemize{
\item Providing three numeric vector will construct points with those x and y
coordinate and a weight.
\item Providing a point and a numeric will construct points with the given weight
}

\strong{3 dimensional weighted points}
\itemize{
\item Providing four numeric vector will construct points with those x, y and z
coordinate and a weight.
\item Providing a point and a numeric will construct points with the given weight
}
}

\examples{
# Construction
wp <- weighted_point(sample(5), sample(5), runif(5))
wp

# Convert to point
p <- as_point(wp)
p

# Construct from point and weight

weighted_point(p, 1:5)

}
