% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{centroid}
\alias{centroid}
\title{Calculate the centroid of a geometry or a set of points}
\usage{
centroid(x, ...)
}
\arguments{
\item{x}{A vector of triangles or tetrahedrons, or points if \code{y}, \code{z}, or \code{t}
are provided}

\item{...}{arguments passed on to methods. Specifically the method for
\code{euclid_geometry} vectors accept up to three additional vectors of
\code{euclid_point} (\code{y}, \code{z}, and \code{t} arguments) if \code{x} is a \code{euclid_point}
vector}
}
\value{
a vector of points
}
\description{
The centroid is the geometric mean of the vertices in a geometry. euclid
supports centroid calculations for triangles, tetrahedrons and 2-4 points
}
\examples{
p <- point(c(1, 5, 6), c(1, -2, 4))
# centroid of triangle
t <- triangle(p[1], p[2], p[3])
centroid(t)

plot(t, col = "grey", border = NA)
euclid_plot(centroid(t), pch = 16, cex = 2, col = "firebrick")

# Same as providing the points directly
centroid(p[1], p[2], p[3])

}
\seealso{
Other Locations: 
\code{\link{barycenter}()},
\code{\link{bisector}()},
\code{\link{circumcenter}()},
\code{\link{equidistant_line}()},
\code{\link{point}()},
\code{\link{radical}()},
\code{\link{weighted_point}()}
}
\concept{Locations}
