% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_locations.R
\name{centroid}
\alias{centroid}
\title{Calculate the centroid of a geometry or a set of points}
\usage{
centroid(x, y = NULL, z = NULL, t = NULL)
}
\arguments{
\item{x}{A vector of triangles or tetrahedrons, or points if \code{y}, \code{z}, or \code{t}
are provided}

\item{y, z, t}{Vector of points}
}
\value{
a vector of points
}
\description{
The centroid is the geometric mean of the vertices in a geometry. euclid
supports centroid calculations for triangles, tetrahedrons and 2-4 points
}
\examples{
# centroid of triangle
t <- triangle(point(1, 1), point(5, -2), point(6, 4))
centroid(t)

# Same as providing the points directly
centroid(point(1, 1), point(5, -2), point(6, 4))

}
