% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\alias{is_circle}
\title{Vector of circles}
\usage{
circle(...)

is_circle(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{x}{A circle vector}
}
\value{
An \code{euclid_circle2} or \code{euclid_circle3} vector
}
\description{
This function create circles from various combinations of input.
}
\section{Constructors}{

\strong{2 dimensional circles}
\itemize{
\item Providing one point and one numeric vector will construct circles centered
at the point with the radius given by the numeric.
\item Providing two point vectors will construct circles centered between the two
points with a radius of half the distance between the two points.
\item Providing three point vectors will construct the unique circle that pass
through the three points.
}
}

\examples{
## 2 Dimensions

point1 <- point(runif(5), runif(5))
point2 <- point(runif(5), runif(5))
point3 <- point(runif(5), runif(5))
number <- exact_numeric(1:5)

# Construction with center and radius
circle(point1, number)

# integers and numerics are converted automatically
circle(point1, 1:5)

# You are free to name the input for readability
circle(center = point1, radius = number)

# Construction with 2 points
circle(point1, point2)

# Construction with 3 points
circle(point1, point2, point3)

}
