% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{compose}
\alias{compose}
\title{Arranges shapes in m on layout in km according to membs vector}
\usage{
compose(m, membs, km, bysize = TRUE, K = 0.05, noverlap = TRUE)
}
\description{
Arranges shapes in m on layout in km according to membs vector
}
