% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_projection.R
\name{normal}
\alias{normal}
\title{Get the normal of a geometry}
\usage{
normal(x, y = NULL)
}
\arguments{
\item{x, y}{Geometry vectors. If \code{y} is given both \code{x} and \code{y} must be
convertible to vectors and be in 3D}
}
\value{
A vector of directions with the same dimensionality as the input
}
\description{
A normal is a direction perpendicular to a geometry. Since normals doesn't
have any magnitude they are returned as directions rather than vectors.
Normals are not given for all geometries. In 2 dimensions they are defined
for line-like geometries (directions, vectors, lines, rays, and segments),
while in 3 dimensions they are given for plane-like geometries (circles,
triangles, and planes).
}
\examples{
# Get normals to a vector of directions
d <- direction(sample(10, 5), sample(10,5))
normal(d)

# Get the normal of a triangle
t <- triangle(point(4, 7, 1), point(9, 2, 1), point(1, 6, 8))
normal(t)

# equivalent to the normal of the supporting plane
normal(t) == normal(as_plane(t))

# Get direction orthogonal to two vectors
normal(vec(5, 2, -7), vec(-1, 4, 9))

}
