% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\alias{is_point}
\alias{as_point}
\title{Vector of points}
\usage{
point(...)

is_point(x)

as_point(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{x}{A point vector or an object to convert to it}
}
\value{
a \code{euclid_point} vector
}
\description{
Points are the fundamental unit in geometry, from which other geometric
primitives can be constructed. Points support less than and greater than
operations which are handled lexicographically. This means that point vectors
can also be sorted and ranked. Subtracking a point from a point gives a
vector. Adding and subtracting a vector to a point translates the point by
the vector.
}
\examples{
num1 <- exact_numeric(runif(5))
num2 <- exact_numeric(runif(5))
num3 <- exact_numeric(runif(5))

# 2 dimensions
point(num1, num2)

# 3 dimensions
point(num1, num2, num3)

# Standard R vectors are automatically converted to exact_numeric vectors
point(runif(5), runif(5))

}
