% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray.R, R/segment.R
\name{ray}
\alias{ray}
\alias{is_ray}
\alias{as_ray}
\alias{is_segment}
\alias{as_segment}
\title{Vector of rays}
\usage{
ray(..., default_dim = 2)

is_ray(x)

as_ray(x)

is_segment(x)

as_segment(x)
}
\arguments{
\item{...}{Various input. See the Constructor section.}

\item{default_dim}{The dimensionality when constructing an empty vector}

\item{x}{A vector of rays or an object to convert to it}
}
\value{
An \code{euclid_ray} vector
}
\description{
A ray is a directed line that starts at a given point and extends to
infinity. As such it does not have a magnitude (like a vector) but only a
beginning and a direction.
}
\section{Constructors}{

\strong{2 and 3 dimensional rays}
\itemize{
\item Providing two points will construct rays starting at the first point and
going through the second
\item Providing a point and a vector will construct rays starting at the point
and moving in the direction of the vector.
}
}

