% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgeometries.R
\name{subgeometries}
\alias{subgeometries}
\alias{vert}
\alias{vert<-}
\alias{edge}
\alias{edge_count}
\title{Extract vertices and edges from geometries}
\usage{
vert(x, which = NULL, ...)

vert(x, which = NULL, ...) <- value

edge(x, which = NULL, ...)

edge_count(x)
}
\arguments{
\item{x}{A vector of geometries}

\item{which}{An integer vector giving the vertex/edge to extract, or \code{NULL}
to extract all.}

\item{...}{arguments passed on to methods}

\item{value}{An \code{euclid_point} vector of the same dimensionality as \code{x}}
}
\value{
A \code{euclid_point} vector for \code{vert()} or a \code{euclid_segment} vector for
\code{edge()} matching the dimensionality of \code{x}
}
\description{
Geometries located in space (all except directions and vectors), have one or
more points supporting it. Those can be extracted and modifed with \code{vert()}.
For geometries with a cardinality above one there's a choice of which support
point to extract/modify or all. Geometries consisting of more than one vertex
(segments, triangles, and tetrahedrons) also have associated edges that can
be extracted (but not modified) with \code{edge()}. \code{edge_count()} provides the
number of edges in each element in the geometry vector. The length of the
output of \code{edge(x)} is \code{sum(edge_count(x))}.
}
\section{Vertex definition}{

For geometries that are defined exclusively by points the definition of the
output is straight forward, e.g. for triangles \code{vert()} will extract one or
all of the corners depending on the value of \code{which}. For the other
geometries the output is defined according to the below:
\itemize{
\item \strong{circles and spheres}: The vertex is the center
\item \strong{rays}: The vertex is the source
\item \strong{lines and planes}: The vertex is an arbitrary point on the geometry
}
}

\examples{
# Get the source vertex in a segment
s <- segment(point(3, 6), point(1, -7))
vert(s, 1)

# And the target
vert(s, 2)

# Not providing an index extracts them all
vert(s)

# Set the source of a segment
vert(s, 1) <- point(0, 0)
s

# Get a point on a line
l <- line(4, 7, -1)
vert(l)

# Setting the vertex of a line moves it so it runs through it
point(1, 2) \%is_on\% l
vert(l) <- point(1, 2)
point(1, 2) \%is_on\% l

# Get one of the sides from a triangle
t <- triangle(point(1, 2), point(6, 3), point(3, 1))
edge(t, 2)

# or all
edge(t)
}
\seealso{
Other Geometry methods: 
\code{\link{def}()},
\code{\link{euclid_geometry}}
}
\concept{Geometry methods}
